veramo_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Pillow==10.1.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.0
django-redis==5.4.0
python-dotenv==1.0.0
reportlab==3.6.13
